refactor(ui): Move repositories view to a sidebar option
In anticipation of the main product page becoming an admin security 
dashboard that shows the top-level security status of the product, move
the earlier view which listed the repository status of the product to
its own sidebar item "Repositories".
