+++ ui/src/routes/_layout/organizations/$orgId/products/$productId/repositories/index.tsx
@@ -23,10 +23,9 @@ import {
   getCoreRowModel,
   useReactTable,
 } from '@tanstack/react-table';
-import { EditIcon } from 'lucide-react';
+import { PlusIcon } from 'lucide-react';
 
 import {
-  useProductsServiceDeleteProductById,
   useProductsServiceGetProductById,
   useRepositoriesServiceGetRepositoriesByProductId,
 } from '@/api/queries';
@@ -34,20 +33,12 @@ import {
   prefetchUseProductsServiceGetProductById,
   prefetchUseRepositoriesServiceGetRepositoriesByProductId,
 } from '@/api/queries/prefetch';
-import { ApiError, Repository } from '@/api/requests';
+import { Repository } from '@/api/requests';
 import { DataTable } from '@/components/data-table/data-table';
-import { DeleteDialog } from '@/components/delete-dialog';
-import { DeleteIconButton } from '@/components/delete-icon-button';
 import { LoadingIndicator } from '@/components/loading-indicator';
 import { ToastError } from '@/components/toast-error';
 import { Button } from '@/components/ui/button';
-import {
-  Card,
-  CardContent,
-  CardDescription,
-  CardHeader,
-  CardTitle,
-} from '@/components/ui/card';
+import { Card, CardContent, CardHeader, CardTitle } from '@/components/ui/card';
 import {
   Tooltip,
   TooltipContent,
@@ -55,10 +46,10 @@ import {
 } from '@/components/ui/tooltip';
 import { toast } from '@/lib/toast';
 import { paginationSearchParameterSchema } from '@/schemas';
-import { LastJobStatus } from './repositories/-components/last-job-status';
-import { LastRunDate } from './repositories/-components/last-run-date';
-import { LastRunStatus } from './repositories/-components/last-run-status';
-import { TotalRuns } from './repositories/-components/total-runs';
+import { LastJobStatus } from './-components/last-job-status';
+import { LastRunDate } from './-components/last-run-date';
+import { LastRunStatus } from './-components/last-run-status';
+import { TotalRuns } from './-components/total-runs';
 
 const defaultPageSize = 10;
 
@@ -102,19 +93,7 @@ const columns = [
     id: 'runs',
     header: 'Runs',
     size: 60,
-    cell: ({ row }) => (
-      <Link
-        to='/organizations/$orgId/products/$productId/repositories/$repoId/runs'
-        params={{
-          orgId: row.original.organizationId.toString(),
-          productId: row.original.productId.toString(),
-          repoId: row.original.id.toString(),
-        }}
-        className='font-semibold text-blue-400 hover:underline'
-      >
-        <TotalRuns repoId={row.original.id} />
-      </Link>
-    ),
+    cell: ({ row }) => <TotalRuns repoId={row.original.id} />,
   }),
   columnHelper.display({
     id: 'runStatus',
@@ -133,9 +112,8 @@ const columns = [
   }),
 ];
 
-const ProductComponent = () => {
+const RepositoriesComponent = () => {
   const params = Route.useParams();
-  const navigate = Route.useNavigate();
   const search = Route.useSearch();
   const pageIndex = search.page ? search.page - 1 : 0;
   const pageSize = search.pageSize ? search.pageSize : defaultPageSize;
@@ -159,35 +137,6 @@ const ProductComponent = () => {
     limit: pageSize,
     offset: pageIndex * pageSize,
   });
-
-  const { mutateAsync: deleteProduct, isPending } =
-    useProductsServiceDeleteProductById({
-      onSuccess() {
-        toast.info('Delete Product', {
-          description: `Product "${product?.name}" deleted successfully.`,
-        });
-        navigate({
-          to: '/organizations/$orgId',
-          params: { orgId: params.orgId },
-        });
-      },
-      onError(error: ApiError) {
-        toast.error(error.message, {
-          description: <ToastError error={error} />,
-          duration: Infinity,
-          cancel: {
-            label: 'Dismiss',
-            onClick: () => {},
-          },
-        });
-      },
-    });
-
-  async function handleDelete() {
-    await deleteProduct({
-      productId: Number.parseInt(params.productId),
-    });
-  }
 
   const table = useReactTable({
     data: repositories?.data || [],
@@ -220,76 +169,54 @@ const ProductComponent = () => {
   }
 
   return (
-    <div className='flex flex-col gap-4'>
-      <Card>
-        <CardHeader>
-          <CardTitle className='flex flex-row justify-between'>
-            <div className='flex items-stretch'>
-              <div className='flex items-center pb-1'>{product.name}</div>
-              <Tooltip>
-                <TooltipTrigger>
-                  <Button
-                    asChild
-                    size='sm'
-                    variant='outline'
-                    className='ml-2 px-2'
-                  >
-                    <Link
-                      to='/organizations/$orgId/products/$productId/edit'
-                      params={{
-                        orgId: params.orgId,
-                        productId: params.productId,
-                      }}
-                    >
-                      <EditIcon className='h-4 w-4' />
-                    </Link>
-                  </Button>
-                </TooltipTrigger>
-                <TooltipContent>Edit this product</TooltipContent>
-              </Tooltip>
-            </div>
-            <DeleteDialog
-              item={{
-                descriptor: 'product',
-                name: product.name,
-              }}
-              onDelete={handleDelete}
-              isPending={isPending}
-              trigger={<DeleteIconButton />}
-            />
-          </CardTitle>
-          <CardDescription>{product.description}</CardDescription>
-        </CardHeader>
-        <CardContent>
-          Here, some security summary information (numbers, icons) could be
-          displayed.
-        </CardContent>
-      </Card>
-      <Card>
-        <CardContent>
-          <DataTable
-            table={table}
-            setCurrentPageOptions={(currentPage) => {
-              return {
-                to: Route.to,
-                search: { ...search, page: currentPage },
-              };
-            }}
-            setPageSizeOptions={(size) => {
-              return {
-                to: Route.to,
-                search: { ...search, page: 1, pageSize: size },
-              };
-            }}
-          />
-        </CardContent>
-      </Card>
-    </div>
+    <Card>
+      <CardHeader>
+        <CardTitle>Repositories belonging to {product.name}</CardTitle>
+        <div className='py-4'>
+          <Tooltip>
+            <TooltipTrigger asChild>
+              <Button asChild size='sm' className='ml-auto gap-1'>
+                <Link
+                  to='/organizations/$orgId/products/$productId/create-repository'
+                  params={{
+                    orgId: params.orgId,
+                    productId: params.productId,
+                  }}
+                >
+                  Add repository
+                  <PlusIcon className='h-4 w-4' />
+                </Link>
+              </Button>
+            </TooltipTrigger>
+            <TooltipContent>
+              Add a repository for managing compliance runs
+            </TooltipContent>
+          </Tooltip>
+        </div>
+      </CardHeader>
+      <CardContent>
+        <DataTable
+          table={table}
+          setCurrentPageOptions={(currentPage) => {
+            return {
+              to: Route.to,
+              search: { ...search, page: currentPage },
+            };
+          }}
+          setPageSizeOptions={(size) => {
+            return {
+              to: Route.to,
+              search: { ...search, page: 1, pageSize: size },
+            };
+          }}
+        />
+      </CardContent>
+    </Card>
   );
 };
 
 export const Route = createFileRoute(
-  '/_layout/organizations/$orgId/products/$productId/'
+  '/_layout/organizations/$orgId/products/$productId/repositories/'
 )({
   validateSearch: paginationSearchParameterSchema,
   loaderDeps: ({ search: { page, pageSize } }) => ({ page, pageSize }),
@@ -308,6 +235,6 @@ export const Route = createFileRoute(
       ),
     ]);
   },
-  component: ProductComponent,
+  component: RepositoriesComponent,
   pendingComponent: LoadingIndicator,
 });
